+++ docs/about/whats_new.rst
@@ -20,6 +20,7 @@ v1.9.next
 - Remove ingestion methods, `GridWorkflow` and `Tile` classes (:pull:`1465`)
 - Fix postgis queries for numeric custom search fields (:pull:`1475`)
 - Document best practice for pulling in changes from develop and update constraints.txt (:pull:`1478`)
+- Postgis index driver performance tuning (:pull:`1480`)
 
 
 v1.8.next